fixing staffServices table after implementing service table

+++ src/validators/staffService.validator.js
@@ -13,23 +13,27 @@ const validatorAssignSerViceToStaff = [
     .notEmpty()
     .withMessage("id is required")
     .isInt({ min: 1 })
-    .withMessage("id must be a positive integer"),
+    .withMessage("id must be a positive integer")
+    .toInt(),
 
   body("service_id")
     .notEmpty()
     .withMessage("id is required")
     .isInt({ min: 1 })
-    .withMessage("id must be a positive integer"),
+    .withMessage("id must be a positive integer")
+    .toInt(),
 
   body("staff_duration_min")
     .optional()
     .isInt({ min: 1 })
-    .withMessage("staff_duration_min must be a positive integer"),
+    .withMessage("staff_duration_min must be a positive integer")
+    .toInt(),
 
   body("staff_price_cents")
     .optional()
     .isInt({ min: 1 })
-    .withMessage("staff_price_cent must be a positive integer"),
+    .withMessage("staff_price_cent must be a positive integer")
+    .toInt(),
   handleVaidationErrors,
 ];
 
@@ -56,13 +60,15 @@ const validatorRemoveServiceFromStaff = [
     .notEmpty()
     .withMessage("staff_id is required")
     .isInt({ min: 1 })
-    .withMessage("staff_id must be a positive integer"),
+    .withMessage("staff_id must be a positive integer")
+    .toInt(),
 
   body("service_id")
     .notEmpty()
     .withMessage("service_id is required")
     .isInt({ min: 1 })
-    .withMessage("service_id must be a positive integer"),
+    .withMessage("service_id must be a positive integer")
+    .toInt(),
   handleVaidationErrors,
 ];
 
@@ -71,23 +77,27 @@ const validatorUpdateStaffService = [
     .notEmpty()
     .withMessage("staff_id is required")
     .isInt({ min: 1 })
-    .withMessage("staff_id must be a positive integer"),
+    .withMessage("staff_id must be a positive integer")
+    .toInt(),
 
   body("service_id")
     .notEmpty()
     .withMessage("service_id is required")
     .isInt({ min: 1 })
-    .withMessage("service_id must be a positive integer"),
+    .withMessage("service_id must be a positive integer")
+    .toInt(),
 
   body("staff_duration_min")
     .optional()
     .isInt({ min: 1 })
-    .withMessage("staff_duration_ min must be a positive integer"),
+    .withMessage("staff_duration_ min must be a positive integer")
+    .toInt(),
 
   body("staff_price_cents")
     .optional()
     .isInt({ min: 1 })
-    .withMessage("staff_price_cents must be a positive integer"),
+    .withMessage("staff_price_cents must be a positive integer")
+    .toInt(),
   body().custom((value) => {
     if (
       value.staff_duration_min === undefined &&
